Make sure a configure script is executable and stamp.h.in exists
When we are creating ./configure from a quilt patch, the resulting file
may not be executable. Either directly from patch(1) or after
dpkg-source is done with it. Make sure it'll be alright when we need
to use it.

`stamp.h.in' is also created by `./Util/preconfig'. It's an empty file,
so we're not creating it using quilt but just touch it here.