Adding autotools related patches
These create ./configure and ./config.h.in. Usually those are generated
using ./Util/preconfig, but for the purpose of building the debian
package, the goal is to avoid calling the preconfig script.

This commit deliberately keeps the patches out of debian/patches/series,
so it doesn't clash with the current .orig tarball which does in fact
ship a ./configure script. After the next release that will not be the
case anymore and ./debian/at2quilt will automatically add the patches to
the series file.