debian/source/options: Ignore Util/difflog.pl, configure and config.h.in for patch generation
This prevents the automatic generation of a debian-changes-$version
patch because of the mentioned files.

configure and config.h.in are changed by us and Util/difflog.pl is in
the upstream git respectively CVS repository, but not in upstream's
distribution tar ball.

See http://www.zsh.org/mla/workers/2007/msg01060.html and
http://zsh.git.sf.net/git/gitweb.cgi?p=zsh/zsh;a=commitdiff;h=9794b45
for details about the exclusion of Util/difflog.pl