Rewrite keyboard handling in zshrc
This change makes proper use of `$terminfo[]', because it makes sure
that zle (zsh's line editor) puts the terminal into application mode
while it is running. Then - and ONLY then - are the values from the
`$terminfo[]' array valid.

While doing that this also does this:

  a) Refactor the actual binding code
  b) Do not bind non-default vi-* widgets to Up and Down.

The change in "b)" may be controversial because it reverts long standing
debian-zshrc behaviour. However, the #zsh channel on freenode gets a
*lot* of people complaining about this from debian and ubuntu systems.
There even was a bug report about this: #383737. Clint commented like
this:

> Unfortunately this boils down to a binary division in a fundamental
> belief, and either way people are unhappy.

I agree.

But I disagree, that the solution is to keep the non-default behaviour
in that case. The package should keep the default zsh behaviour with
things that boil down to "fundamental belief" and leave it to the user
to decide.

The bindings that were the most controversial were `Up' and `Down' bound
to vi-up/down-line-or-history in viins and vicmd. This patch binds both
keys to the less surprising up/down-line-or-history bindings.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>